pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "phenoradar"
version = "0.1.0"
description = "Macroevolutionary phenotype predictor"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Tomoya Nishiguchi", email = "t.nishiguchi5@gmail.com" }
]
dependencies = [
  "typer>=0.12.5,<1.0.0",
  "polars>=1.9.0,<2.0.0",
  "scikit-learn>=1.5.2,<2.0.0",
  "matplotlib>=3.9.2,<4.0.0",
  "optuna>=4.0.0,<5.0.0",
  "PyYAML>=6.0.2,<7.0.0",
  "pydantic>=2.9.2,<3.0.0",
  "joblib>=1.4.2,<2.0.0",
]

[project.scripts]
phenoradar = "phenoradar.launcher:main"
phrad = "phenoradar.launcher:main"

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9,<1.0.0",
  "mypy>=1.11.2,<2.0.0",
  "pytest>=8.3.3,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "pre-commit>=3.8.0,<4.0.0",
  "types-PyYAML>=6.0.12.20250915,<7.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/phenoradar"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["sklearn", "sklearn.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["scipy", "scipy.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["joblib", "joblib.*"]
ignore_missing_imports = true
